// src/types/bindings/Customer.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CustomerType } from "./CustomerType";

export type Customer = { id: string, type: CustomerType, name: string, contact_person: string | null, email: string | null, phone: string | null, address: string | null, ssm_no: string | null, nric: string | null, tax_no: string | null, notes: string | null, archived: boolean, created_at: string, updated_at: string, };