// src/types/bindings/QuotationWithLines.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Quotation } from "./Quotation";
import type { QuotationLineItem } from "./QuotationLineItem";

export type QuotationWithLines = { quotation: Quotation, lines: Array<QuotationLineItem>, };